requirements.txt
streamlit
openai
pandas
python-dotenv
plotly
delta-spark
pyspark
pyarrow
